Additionally check reputation colour for IP
The reputation colour and the statistics that come with it is one of the
few tools that SNDS/Microsoft offers to third-party email providers to
figure out when they might be having trouble with sending email to them.

This work was done by Kienan Stewart. This commit is just a squashing of
the work found in the Koumbit/check_snds repository into the single
check_snds script.